Refactors source data fns into single file
