+++ src/app/(site)/shelf/page.tsx
@@ -1,7 +1,7 @@
 import type { Metadata } from "next";
 import cx from "classnames";
 import Book from "src/components/source/book";
-import getAllBooks from "src/data/getAllBooks";
+import { getAllBooks } from "src/data/source";
 import listStyles from "src/styles/list.module.css";
 import libraryStyles from "src/styles/library.module.css";
 
@@ -18,7 +18,7 @@ export default async function Library() {
         return (
           <li key={book._id}>
             {/* @ts-expect-error Server Component */}
-            <Book {...book} link size="medium" />
+            <Book {...book} link size='medium' />
           </li>
         );
       })}
