+++ src/data/source.ts
@@ -1,0 +1,88 @@
+import { groq } from "next-sanity";
+import { SanityDocument } from "@sanity/client";
+import Sanity from "src/providers/sanity";
+import { Note } from "src/types/note";
+import { Source, SourceType } from "src/types/source";
+
+export interface CreateSourceArgs {
+  type: SourceType;
+  title: string;
+  subtitle?: string;
+  author?: string;
+  url?: string;
+  isbn?: string;
+}
+
+export async function createSource(
+  args: CreateSourceArgs
+): Promise<SanityDocument<Source>> {
+  const source = await Sanity.create({
+    _type: "source",
+    ...args,
+  });
+  return source;
+}
+
+export async function getSource(id: string): Promise<Source> {
+  const GET_SOURCE = groq`*[_type == "source" && _id == "${id}"][0] {
+    _id,
+    _createdAt,
+    type,
+    title,
+    subtitle,
+    author,
+    "imageUrl": cover.asset->url,
+    url,
+    isbn
+  }
+  `;
+  return Sanity.fetch<Source>(GET_SOURCE);
+}
+
+export async function getSourceNotes(id: string): Promise<Note[]> {
+  const GET_SOURCE_NOTES = groq`*[_type=='note' && references('${id}')] {
+    _id,
+    _createdAt,
+    body,
+    page,
+    source -> {
+      _id,
+      type,
+      title,
+      url
+    }
+  }`;
+  return Sanity.fetch<Note[]>(GET_SOURCE_NOTES);
+}
+
+export async function getAllSources() {
+  const GET_ALL_SOURCES = groq`*[_type == "source"] | order(_createdAt desc){
+    _id,
+    _createdAt,
+    type,
+    title,
+    subtitle,
+    author,
+    "imageUrl": cover.asset->url,
+    url,
+    isbn
+  }
+  `;
+  return Sanity.fetch<Source[]>(GET_ALL_SOURCES);
+}
+
+export async function getAllBooks() {
+  const GET_ALL_BOOKS = groq`*[_type == "source" && type == "book"] | order(_createdAt desc){
+    _id,
+    _createdAt,
+    type,
+    title,
+    subtitle,
+    author,
+    "imageUrl": cover.asset->url,
+    url,
+    isbn
+  }
+  `;
+  return Sanity.fetch<Source[]>(GET_ALL_BOOKS);
+}
